This is essentially a rewrite of the POS (that does not stand for "Part Of Speech" :D) that I dumped in the repo the other day.
This is a viewer for the Apertium pipeline. You feed it a modes file when you start it up. Then you paste stuff. You can fiddle with any of the stages and see it magically update the remaining stages.

Remember to update config.py for you setup.

I will soon move all of these files into their own directory.

git-svn-id: https://apertium.svn.sourceforge.net/svnroot/apertium@2296 72bbbca6-d526-0410-a7d9-f06f51895060